Added Cards to Feed, bunch of minor stuff
Note that some Images may not load on Web-Target due to the way Canvas fetches the images, quite likely resulting in an issue with CORS for certain URLs. Alternatively <img> loading could be used, but this only makes sense if many images have to be displayed simultaneously - so sticking to Canvas and changing the imageUrls serverside is a better choice regarding reliable performance metrics.